refactor serialisation of device information to support unity
